requirements.txt
# ── Core LangGraph / LangChain ────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# ── LLM providers ────────────────────────────────────────────────────────────
openai>=1.40.0

# ── Vector databases ──────────────────────────────────────────────────────────
chromadb>=0.5.0
faiss-cpu>=1.8.0          # swap for faiss-gpu on GPU instances

# ── API framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── Web search (optional – install the one you use) ───────────────────────────
tavily-python>=0.3.0      # recommended
# google-search-results   # SerpAPI alternative

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.3.0
tiktoken>=0.7.0

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0             # for FastAPI test client
